requirements.txt
boto3==1.35.91
browser-cookie3==0.20.1
faster-whisper==1.1.0
flatbuffers==25.9.23
flask==3.1.0
groq==0.15.0
lz4==4.4.4
mpmath==1.3.0
openai==1.58.1
pycryptodomex==3.23.0
python-dotenv==1.0.1
pywin32==310
requests==2.32.3
shadowcopy==0.0.4
sympy==1.14.0
typing_extensions==4.15.0
urllib3==2.5.0
WMI==1.5.1
yt-dlp==2025.10.14